Added rooms header file.
This is a library for the main Rooms.cpp. I didn't want to clutter the main file like I did with my old Dice game. This will handle most of the basic game mechanics and operations. It will change over time and is nowhere near finished. I am currently new to file stream I/O, so this will take some practice and a lot of trial and error.